Move prometheus measurement to common package
